handle view map from exterior

+++ includes/color_utils.h
@@ -6,7 +6,7 @@
 /*   By: jberredj <jberredj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/06 11:59:19 by jberredj          #+#    #+#             */
-/*   Updated: 2021/12/26 02:00:06 by jberredj         ###   ########.fr       */
+/*   Updated: 2022/01/21 14:23:50 by jberredj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define COLOR_UTILS_H
 # include <stdint.h>
 
-int	add_shade(int color, double shade);
+int	add_shade(int color, float shade);
 int	argb(int a, int r, int g, int b);
 int	blend_argb(int colora, int colorb);
 int	get_a(int trgb);
