+++ libs/libft/srcs/includes/ft_to.h
@@ -6,7 +6,7 @@
 /*   By: jberredj <jberredj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/22 14:25:08 by jberredj          #+#    #+#             */
-/*   Updated: 2021/07/21 22:07:58 by jberredj         ###   ########.fr       */
+/*   Updated: 2022/01/21 14:23:50 by jberredj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ char		*ft_ulltoa_base(unsigned long long n, char *base);
 size_t		ft_ulllen(unsigned long long n);
 size_t		ft_ulllen_base(unsigned long long n, int base);
 uint32_t	ft_atoui32_base(const char *str, char *base);
-double		ft_atof(const char *str);
+float		ft_atof(const char *str);
 int			ft_tolower(int c);
 int			ft_toupper(int c);
 #endif
