+++ srcs/game_engine/rays/utils.c
@@ -6,7 +6,7 @@
 /*   By: jberredj <jberredj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/05 15:33:37 by jberredj          #+#    #+#             */
-/*   Updated: 2021/12/26 01:26:23 by jberredj         ###   ########.fr       */
+/*   Updated: 2022/01/21 14:23:50 by jberredj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 #include "structs/t_ray.h"
 #include "player.h"
 
-double	normalize_angle(double angle)
+float	normalize_angle(float angle)
 {
 	angle = fmod(angle, (2 * M_PI));
 	if (angle < 0)
